requirements.txt
streamlit==1.31.0
opencv-python-headless==4.9.0.80
numpy==1.26.3
pandas==2.2.0
Pillow==10.2.0